Add files via upload
Reproducible Research Course 📓 ♻️
This repository contains the two peer-graded projects necessary to complete the Reproducible Research Course by Johns Hopkins University at Coursera, which is the 5th out of 10 courses in the Data Science Specialization.

From the course homepage:

"This course focuses on the concepts and tools behind reporting modern data analyses in a reproducible manner. Reproducible research is the idea that data analyses, and more generally, scientific claims, are published with their data and software code so that others may verify the findings and build upon them. This course will focus on literate statistical analysis tools which allow one to publish data analyses in a single document that allows others to easily execute the same analysis to obtain the same results." (About this course, Reproducible Research by Johns Hopkins University, September 2017)

There are two projects in this repository:

Course Project 1. Analyzing Data from Personal Activity Monitors

Course Project2. Exploring the National Oceanic and Atmospheric Administration's (NOAA) storm database

Each of the project folders is, of course, reproducible; this means that they are self-contained: all code + data (or link to data) necessary to produce the reports.